Consolidated the auth package into a common location
