+++ cmd/registry/main.go
@@ -22,7 +22,7 @@ import (
 	"github.com/codegangsta/cli"
 
 	"github.com/amalgam8/amalgam8/registry/api"
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/cluster"
 	"github.com/amalgam8/amalgam8/registry/config"
 	"github.com/amalgam8/amalgam8/registry/replication"
