+++ controller/api/util.go
@@ -17,7 +17,7 @@ package api
 import (
 	"time"
 
-	"github.com/amalgam8/amalgam8/controller/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/controller/metrics"
 	"github.com/amalgam8/amalgam8/controller/util"
 	"github.com/ant0ine/go-json-rest/rest"
