+++ pkg/auth/errors.go
