+++ pkg/auth/trusted.go
