+++ pkg/auth/trusted_test.go
