+++ registry/api/config.go
@@ -17,7 +17,7 @@ package api
 import (
 	"github.com/ant0ine/go-json-rest/rest"
 
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/store"
 )
