+++ registry/api/middleware/accesslog.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/amalgam8/amalgam8/registry/api/env"
 	"github.com/amalgam8/amalgam8/registry/api/protocol"
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/utils/logging"
 )
