+++ registry/api/middleware/auth.go
@@ -21,7 +21,7 @@ import (
 	"github.com/ant0ine/go-json-rest/rest"
 
 	"github.com/amalgam8/amalgam8/registry/api/env"
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/utils/i18n"
 )
