+++ registry/api/server_test.go
@@ -31,7 +31,7 @@ import (
 
 	"github.com/amalgam8/amalgam8/registry/api/protocol/amalgam8"
 	"github.com/amalgam8/amalgam8/registry/api/uptime"
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/store"
 	"github.com/amalgam8/amalgam8/registry/utils/health"
 )
