--- registry/auth/default.go
@@ -1,33 +1,0 @@
-// Copyright 2016 IBM Corporation
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License.
-
-package auth
-
-type defaultAuthenticator struct{}
-
-var defaultAuth = &defaultAuthenticator{}
-var defaultNamespace = Namespace("default")
-
-// DefaultAuthenticator returns the default authenticator provided by this package,
-// which resolves "" (empty string) as well as "default" token to the default namespace.
-func DefaultAuthenticator() Authenticator {
-	return defaultAuth
-}
-
-func (aut *defaultAuthenticator) Authenticate(token string) (*Namespace, error) {
-	if token == "" || token == defaultNamespace.String() {
-		return &defaultNamespace, nil
-	}
-	return nil, ErrUnrecognizedToken
-}
