--- registry/auth/namespace.go
@@ -1,28 +1,0 @@
-// Copyright 2016 IBM Corporation
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License.
-
-package auth
-
-// Namespace represents the service registration and discovery scope
-type Namespace string
-
-// NamespaceFrom returns the namespace identified by the given string
-func NamespaceFrom(s string) Namespace {
-	return Namespace(s)
-}
-
-// String returns the string representation of the given Namespace
-func (ns Namespace) String() string {
-	return string(ns)
-}
