+++ registry/replication/replication_message.go
@@ -17,7 +17,7 @@ package replication
 import (
 	"fmt"
 
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/cluster"
 )
