+++ registry/replication/replicator.go
@@ -15,7 +15,7 @@
 package replication
 
 import (
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/cluster"
 	"github.com/amalgam8/amalgam8/registry/utils/channels"
 )
