+++ registry/replication/server.go
@@ -28,7 +28,7 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/cluster"
 	"github.com/amalgam8/amalgam8/registry/utils/channels"
 	"github.com/amalgam8/amalgam8/registry/utils/health"
