+++ registry/store/filesystem/factory.go
@@ -19,7 +19,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/store"
 )
