+++ registry/store/kubernetes/catalog.go
@@ -22,7 +22,7 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/store"
 	"github.com/amalgam8/amalgam8/registry/utils/logging"
 )
