+++ registry/store/kubernetes/factory.go
@@ -15,7 +15,7 @@
 package kubernetes
 
 import (
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/store"
 )
