+++ registry/store/replicated_catalog.go
@@ -20,7 +20,7 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 
-	"github.com/amalgam8/amalgam8/registry/auth"
+	"github.com/amalgam8/amalgam8/pkg/auth"
 	"github.com/amalgam8/amalgam8/registry/replication"
 	"github.com/amalgam8/amalgam8/registry/utils/channels"
 	"github.com/amalgam8/amalgam8/registry/utils/logging"
